[IMP] l10n_my_*: merge modules
In 4424a0f, we added a new module for the Malaysian
EDI in order to bring some missing features to the
EDI.
This commit will merge that new module with the
original one, keeping all the features in a single
place.

task-4516544
