+++ addons/l10n_my_edi/__manifest__.py
@@ -15,14 +15,18 @@
         'data/ir_cron.xml',
         'data/l10n_my_edi.industry_classification.csv',
         'data/my_ubl_templates.xml',
+
         'security/ir.model.access.csv',
+
         'views/account_move_view.xml',
         'views/account_tax_view.xml',
         'views/l10n_my_edi_industrial_classification_views.xml',
         'views/product_template_view.xml',
+        'views/report_invoice.xml',
         'views/res_company_view.xml',
         'views/res_config_settings_view.xml',
         'views/res_partner_view.xml',
+
         'wizard/l10n_my_edi_status_update_wizard.xml',
     ],
     'installable': True,
