+++ addons/l10n_my_edi/models/__init__.py
@@ -2,6 +2,7 @@
 from . import account_edi_proxy_user
 from . import account_edi_xml_ubl_my
 from . import account_move
+from . import account_move_line
 from . import account_move_send
 from . import account_tax
 from . import l10n_my_edi_industry_classification
