+++ addons/l10n_my_edi/models/res_company.py
@@ -6,13 +6,6 @@ from odoo import api, fields, models
 class ResCompany(models.Model):
     _inherit = 'res.company'
 
-    # ---------------
-    # Default methods
-    # ---------------
-
-    def _default_l10n_my_edi_industrial_classification(self):
-        return self.env.ref('l10n_my_edi.class_00000', raise_if_not_found=False)
-
     # ------------------
     # Fields declaration
     # ------------------
@@ -24,11 +17,7 @@ class ResCompany(models.Model):
     l10n_my_identification_type = fields.Selection(related='partner_id.l10n_my_identification_type', readonly=False)
     l10n_my_identification_number = fields.Char(related='partner_id.l10n_my_identification_number', readonly=False)
     l10n_my_identification_number_placeholder = fields.Char(compute="_compute_l10n_my_identification_number_placeholder")
-    l10n_my_edi_industrial_classification = fields.Many2one(
-        comodel_name='l10n_my_edi.industry_classification',
-        string="Ind. Classification",
-        default=_default_l10n_my_edi_industrial_classification,
-    )
+    l10n_my_edi_industrial_classification = fields.Many2one(related='partner_id.l10n_my_edi_industrial_classification', readonly=False)
     l10n_my_edi_mode = fields.Selection(
         selection=[
             ('test', 'Pre-Production'),
