+++ addons/l10n_my_edi/models/res_partner.py
@@ -41,12 +41,24 @@ class ResPartner(models.Model):
     )
     l10n_my_identification_number = fields.Char(string="ID Number")
     l10n_my_identification_number_placeholder = fields.Char(compute="_compute_l10n_my_identification_number_placeholder")
+    l10n_my_edi_industrial_classification = fields.Many2one(
+        comodel_name='l10n_my_edi.industry_classification',
+        string="Ind. Classification",
+        compute='_compute_l10n_my_edi_industrial_classification',
+        store=True,
+        readonly=False,
+    )
+    l10n_my_edi_malaysian_tin = fields.Char(
+        string="Malaysian TIN",
+        help="The value set in this field will be used as TIN for the customer/supplier.\n"
+             "If left empty, the Tax ID field will be used.",
+    )
 
     # --------------------------------
     # Compute, inverse, search methods
     # --------------------------------
 
-    @api.depends('l10n_my_identification_type', 'l10n_my_identification_number', 'vat')
+    @api.depends('l10n_my_identification_type', 'l10n_my_identification_number', 'vat', 'l10n_my_edi_malaysian_tin')
     def _compute_l10n_my_tin_validation_state(self):
         """ The three @depends are used for the validation. If they change, we will invalidate it and expect the user to revalidate. """
         self.l10n_my_tin_validation_state = False
@@ -77,6 +89,10 @@ class ResPartner(models.Model):
             elif partner.l10n_my_identification_type == 'ARMY':
                 placeholder = '830805-13-4983'
             partner.l10n_my_identification_number_placeholder = placeholder
+
+    def _compute_l10n_my_edi_industrial_classification(self):
+        default_classification = self.env.ref('l10n_my_edi.class_00000', raise_if_not_found=False)
+        self.filtered(lambda p: not p.l10n_my_edi_industrial_classification).l10n_my_edi_industrial_classification = default_classification
 
     # --------------
     # Action methods
@@ -115,8 +131,11 @@ class ResPartner(models.Model):
     def _l10n_my_edi_get_tin_for_myinvois(self):
         """ Helper to return the VAT number relevant to the situation. """
         self.ensure_one()
-        return self.vat
+        # When l10n_my_edi_malaysian_tin is set, it will be used instead of the VAT.
+        # A user may want to keep the correct VAT on a foreign contact while also use myinvois with a malaysia TIN/Generic TIN
+        # Using the Tax ID field also causes issue when base_vat is enabled, which block setting foreign VAT numbers.
+        return self.l10n_my_edi_malaysian_tin or self.vat
 
     @api.model
     def _commercial_fields(self):
-        return super()._commercial_fields() + ['l10n_my_identification_type', 'l10n_my_identification_number']
+        return super()._commercial_fields() + ['l10n_my_identification_type', 'l10n_my_identification_number', 'l10n_my_edi_industrial_classification', 'l10n_my_edi_malaysian_tin']
