+++ addons/l10n_my_edi/views/res_company_view.xml
@@ -16,7 +16,7 @@
                                 <field name="l10n_my_identification_number" options="{'placeholder_field': 'l10n_my_identification_number_placeholder'}"/>
                             </span>
                         </div>
-                        <field class="w-50" name="l10n_my_edi_industrial_classification"/>
+                        <field class="w-50" name="l10n_my_edi_industrial_classification" options='{"no_open": True}'/>
                     </group>
                 </group>
             </xpath>
