--- addons/l10n_my_edi_extended/__init__.py
@@ -1,2 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-from . import models
