--- addons/l10n_my_edi_extended/models/__init__.py
@@ -1,6 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-from . import account_edi_xml_ubl_my
-from . import account_move
-from . import account_move_line
-from . import account_move_send
-from . import res_partner
