--- addons/l10n_my_edi_extended/tests/__init__.py
@@ -1,3 +1,0 @@
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-from . import test_file_generation
-from . import test_new_submissions_flow
