--- addons/l10n_my_edi_extended/views/res_partner_view.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <record id="view_partner_form_inherit_l10n_my_myinvois_extended" model="ir.ui.view">
-        <field name="name">res.partner.form.inherit.l10n_my_myinvois_extended</field>
-        <field name="model">res.partner</field>
-        <field name="inherit_id" ref="l10n_my_edi.view_partner_form_inherit_l10n_my_myinvois"/>
-        <field name="arch" type="xml">
-            <xpath expr="//group[@name='l10n_my_edi']/group" position="inside">
-                <field name="l10n_my_edi_industrial_classification" readonly="parent_id"/>
-                <field name="l10n_my_edi_malaysian_tin" placeholder="EI00000000020" readonly="parent_id"/>
-            </xpath>
-        </field>
-    </record>
-</odoo>
